refactor(file structure): move modules/domAction.mjs to features/reading-list/action.js
