+++ features/reading-list/action.js
@@ -1,4 +1,4 @@
-import * as extension from './chrome/runtime.mjs'
+import * as extension from '../../modules/chrome/runtime.mjs'
 
 const activeLi = () => $('.active')
 const activeUrl = () => activeLi().find('a').attr('href')
